requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
orjson==3.10.7
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.1.1
chromadb==0.5.5
httpx==0.27.2
beautifulsoup4==4.12.3
streamlit==1.39.0
faiss-cpu==1.8.0.post1
openai==1.51.2
python-dotenv==1.0.1
